requirements.txt
argcomplete==0.8.4
beautifulsoup4==4.3.2
blinker==1.3
chardet==2.3.0
EbookLib==0.15
Flask==0.10.1
Flask-Bootstrap==3.3.0.1
Flask-Login==0.2.11
Flask-Mail==0.9.1
Flask-Markdown==0.3
flask-mongoengine==0.7.1
Flask-PageDown==0.1.5
Flask-Principal==0.4.0
Flask-Security==1.7.4
Flask-WTF==0.11
gunicorn==19.3.0
html5lib==0.999
itsdangerous==0.24
Jinja2==2.7.3
lxml==3.4.2
Markdown==2.6
MarkupSafe==0.23
git+https://github.com/MongoEngine/mongoengine.git@7d1ba466
passlib==1.6.2
pdfminer==20140328
Pillow==2.7.0
pymongo==2.8
pypandoc==0.9.6
PyPDF2==1.24
python-docx==0.8.5
python-pptx==0.5.7
reportlab==3.1.44
six==1.9.0
SpeechRecognition==1.2.0
textract==1.2.0
Werkzeug==0.10.1
WTForms==2.0.2
xhtml2pdf==0.0.6
xlrd==0.9.3
XlsxWriter==0.7.2
